requirements.txt
# ─── API ─────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── UI ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ─── Transcription ───────────────────────────────────────────────────────────
groq>=0.9.0

# ─── Embeddings + segmentation ───────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── Audio ───────────────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.0

# ─── LLM scoring (DeepSeek / GLM via NVIDIA NIM) ─────────────────────────────
openai>=1.30.0

# ─── Video processing ────────────────────────────────────────────────────────
# IMPORTANT: install opencv-python BEFORE scenedetect
opencv-python>=4.9.0
scenedetect[opencv]>=0.6.3

# ─── TensorFlow CPU (MoViNet) — install separately after the above ────────────
# pip install tensorflow-cpu>=2.16.0  tensorflow-hub>=0.16.0
# (large download ~700 MB — run pip install tensorflow-cpu tensorflow-hub)

# ─── Queue (Celery + Redis) ───────────────────────────────────────────────────
# Upstash free tier: rediss://:password@hostname.upstash.io:6380/0
celery[redis]>=5.4.0
# redis>=5.0.0  # pulled in by celery[redis]

# ─── Database ─────────────────────────────────────────────────────────────────
# Supabase / any Postgres: set DATABASE_BACKEND=postgres + DATABASE_URL
psycopg2-binary>=2.9.9

# ─── Data / utils ────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.7.0
requests>=2.32.0
python-dotenv>=1.0.0
